tests: Move all the test code in vector_selection_test.cc
