+++ cpp/src/arrow/compute/kernels/CMakeLists.txt
@@ -94,7 +94,6 @@ add_arrow_compute_test(vector_test
                        vector_nested_test.cc
                        vector_replace_test.cc
                        vector_run_end_encode_test.cc
-                       vector_run_end_selection_test.cc
                        select_k_test.cc
                        EXTRA_LINK_LIBS
                        arrow_compute_kernels_testing)
